+++ server/src/entity/dynamic_monitoring_summary.rs
@@ -6,7 +6,7 @@ use serde::{Deserialize, Serialize};
 pub struct Model {
     #[sea_orm(primary_key)]
     pub id: i64,
-    pub uuid_id: i16,
+    pub uuid_id: i32,
     pub timestamp: i64,
     pub cpu_usage: Option<i16>,
     pub gpu_usage: Option<i16>,
