+++ server/src/rpc/agent/query_dynamic_multi_last.rs
@@ -85,7 +85,7 @@ pub async fn dynamic_data_multi_last_query(
         let uuid_cache = MonitoringUuidCache::global();
 
         // Resolve UUIDs to uuid_ids
-        let mut uuid_id_pairs: Vec<(Uuid, i16)> = Vec::with_capacity(deduped_uuids.len());
+        let mut uuid_id_pairs: Vec<(Uuid, i32)> = Vec::with_capacity(deduped_uuids.len());
         for uuid in &deduped_uuids {
             let uuid_id = uuid_cache.get_id(uuid).await.ok_or_else(|| {
                 NodegetError::NotFound(format!(
@@ -146,7 +146,7 @@ fn dedupe_uuids(uuids: Vec<Uuid>) -> Vec<Uuid> {
 }
 
 fn build_union_last_statement(
-    uuid_id_pairs: &[(Uuid, i16)],
+    uuid_id_pairs: &[(Uuid, i32)],
     fields: &[DynamicDataQueryField],
     db: &DatabaseConnection,
 ) -> anyhow::Result<Statement> {
@@ -166,7 +166,7 @@ fn build_union_last_statement(
     ))
 }
 
-fn build_single_last_select(uuid_id: i16, fields: &[DynamicDataQueryField]) -> SelectStatement {
+fn build_single_last_select(uuid_id: i32, fields: &[DynamicDataQueryField]) -> SelectStatement {
     let inner_query = dynamic_monitoring::Entity::find()
         .select_only()
         .column(dynamic_monitoring::Column::UuidId)
@@ -255,7 +255,7 @@ async fn execute_statement_query(
                     // Translate uuid_id → uuid string
                     if let Some(uuid_id_val) = obj.remove("uuid_id") {
                         if let Some(uuid_id) = uuid_id_val.as_i64() {
-                            if let Some(uuid) = uuid_cache.get_uuid(uuid_id as i16).await {
+                            if let Some(uuid) = uuid_cache.get_uuid(uuid_id as i32).await {
                                 obj.insert(
                                     "uuid".to_owned(),
                                     serde_json::Value::String(uuid.to_string()),
